+++ src/Command/SyncCommand.php
@@ -22,7 +22,7 @@ use Webf\FlysystemFailoverBundle\Event\SyncService\SearchingFilesToReplicateStar
 use Webf\FlysystemFailoverBundle\Flysystem\FailoverAdaptersLocatorInterface;
 use Webf\FlysystemFailoverBundle\Service\SyncService;
 
-class SyncCommand extends Command
+final class SyncCommand extends Command
 {
     public function __construct(
         private EventDispatcherInterface $eventDispatcher,
@@ -32,6 +32,7 @@ class SyncCommand extends Command
         parent::__construct();
     }
 
+    #[\Override]
     protected function execute(
         InputInterface $input,
         OutputInterface $output,
@@ -48,7 +49,7 @@ class SyncCommand extends Command
 
         $io = new SymfonyStyle($input, $output);
 
-        $stats = new class() {
+        $stats = new class {
             /** @var array<int, int> */
             private array $nbReplicatedMap = [];
 
@@ -143,7 +144,7 @@ class SyncCommand extends Command
                 $message = $event->getMessage();
 
                 $io->write(sprintf(
-                    'Dispatching message to replicate file ' .
+                    'Dispatching message to replicate file '.
                     '<comment>%s</comment> from storage %s to %s...',
                     $message->getPath(),
                     $message->getInnerSourceAdapter(),
@@ -165,7 +166,7 @@ class SyncCommand extends Command
             DeleteFileMessagePreDispatch::class,
             function (DeleteFileMessagePreDispatch $event) use ($io, $stats) {
                 $io->write(sprintf(
-                    'Dispatching message to delete file ' .
+                    'Dispatching message to delete file '.
                     '<comment>%s</comment> from storage %s...',
                     $event->getMessage()->getPath(),
                     $event->getMessage()->getInnerDestinationAdapter(),
@@ -219,12 +220,13 @@ class SyncCommand extends Command
         return 0;
     }
 
+    #[\Override]
     protected function configure(): void
     {
         $this
             ->setName('webf:flysystem-failover:sync')
             ->setDescription(
-                'Synchronize storages to replicate all files present in the ' .
+                'Synchronize storages to replicate all files present in the '.
                 'first one to the others'
             )
         ;
@@ -237,8 +239,8 @@ class SyncCommand extends Command
             $this->addArgument(
                 'adapter',
                 InputArgument::REQUIRED,
-                'Name of the failover adapter for which to scan the ' .
-                'underlaying storages' .
+                'Name of the failover adapter for which to scan the '.
+                'underlaying storages'.
                 sprintf(
                     ' (one of <comment>"%s"</comment>)',
                     join('"</comment>, <comment>"', $adapters)
@@ -250,7 +252,7 @@ class SyncCommand extends Command
             'extra-files',
             mode: InputOption::VALUE_OPTIONAL,
             description: sprintf(
-                'How to handle extra files in secondary storages. One of ' .
+                'How to handle extra files in secondary storages. One of '.
                 '<comment>"%s"</comment>.',
                 join('"</comment>, <comment>"', [
                     SyncService::EXTRA_FILES_IGNORE,
@@ -264,7 +266,7 @@ class SyncCommand extends Command
         $this->addOption(
             'ignore-modification-dates',
             mode: InputOption::VALUE_NONE,
-            description: 'Do not replicate files already present in ' .
+            description: 'Do not replicate files already present in '.
                 'secondary storages, even if the modification date is older.'
         );
     }