+++ src/Event/SyncService/ListingContentFailed.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace Webf\FlysystemFailoverBundle\Event\SyncService;
 
-class ListingContentFailed extends AbstractListingContentEvent
+final class ListingContentFailed extends AbstractListingContentEvent
 {
 }