+++ src/Event/SyncService/SearchingFilesToReplicateStarted.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Webf\FlysystemFailoverBundle\Event\SyncService;
 
-class SearchingFilesToReplicateStarted
+final class SearchingFilesToReplicateStarted
 {
     public function __construct(
         private string $failoverAdapter,