+++ src/EventListener/DoctrineSchemaListener.php
@@ -1,20 +1,13 @@
 <?php
 
-/*
- * This file is part of the Symfony package.
- *
- * (c) Fabien Potencier <fabien@symfony.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
+declare(strict_types=1);
 
 namespace Webf\FlysystemFailoverBundle\EventListener;
 
 use Doctrine\ORM\Tools\Event\GenerateSchemaEventArgs;
 use Webf\FlysystemFailoverBundle\MessageRepository\DoctrineMessageRepository;
 
-class DoctrineSchemaListener
+final class DoctrineSchemaListener
 {
     public function __construct(private DoctrineMessageRepository $repository)
     {