+++ src/Exception/FailoverAdapterNotFoundException.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Webf\FlysystemFailoverBundle\Exception;
 
-class FailoverAdapterNotFoundException extends OutOfBoundsException
+final class FailoverAdapterNotFoundException extends OutOfBoundsException
 {
     public static function withName(string $name): FailoverAdapterNotFoundException
     {