+++ src/Message/ReplicateFile.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Webf\FlysystemFailoverBundle\Message;
 
-class ReplicateFile implements MessageInterface
+final class ReplicateFile implements MessageInterface
 {
     public function __construct(
         private string $failoverAdapter,
@@ -15,31 +15,37 @@ class ReplicateFile implements MessageInterface
     ) {
     }
 
+    #[\Override]
     public function getFailoverAdapter(): string
     {
         return $this->failoverAdapter;
     }
 
+    #[\Override]
     public function getPath(): string
     {
         return $this->path;
     }
 
+    #[\Override]
     public function getInnerSourceAdapter(): int
     {
         return $this->innerSourceAdapter;
     }
 
+    #[\Override]
     public function getInnerDestinationAdapter(): int
     {
         return $this->innerDestinationAdapter;
     }
 
+    #[\Override]
     public function getRetryCount(): int
     {
         return $this->retryCount;
     }
 
+    #[\Override]
     public function __toString()
     {
         return sprintf(