+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## v0.7.0 (unreleased)
+
+### 💥 Breaking changes
+
+* Make classes final
+
+### ✨ New features
+
+* Implement `ChecksumProvider`, `PublicUrlGenerator` and `TemporaryUrlGenerator`
+
 ## v0.6.0 (February 15, 2025)
 
 ### 💥 Breaking changes